+++ files/RELEASES.md
@@ -110,3 +110,13 @@ Based on:
 - [csharp v4.1.0] files
 ### Releases
 - [NuGet v4.1.0] https://www.nuget.org/packages/WingspanFiles/4.1.0 - files
+
+## 2024-02-19 01:24:02
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.183.2 (2.262.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.1.1] files
+### Releases
+- [NuGet v4.1.1] https://www.nuget.org/packages/WingspanFiles/4.1.1 - files